fix: refactor tests folder structure
