+++ test/productService/productsList.test.js
@@ -1,12 +1,12 @@
-import { handler } from "../src/productService/productsList";
+import { handler } from "../../src/productService/productsList";
 
 const mockedProducts = [{ id: 1 }, { id: 2 }];
 
-jest.mock("../src/utils/envUtils", () => ({
+jest.mock("../../src/utils/envUtils", () => ({
   checkRequiredEnvValues: jest.fn(),
 }));
 
-jest.mock("../src/productService/productModule/productRepository", function () {
+jest.mock("../../src/productService/productModule/productRepository", function () {
   return () => ({
     getAllProducts: jest.fn(() => mockedProducts),
   });
